# First try of using parallelism in the sort operation