--- a/src/global/Constants.h
+++ b/src/global/Constants.h
@@ -75,3 +75,10 @@ static constexpr uint8_t NUM_COMPRESSION_PREFIXES = 127;
 // compression has been applied to  a word
 static const uint8_t NO_PREFIX_CHAR =
     MIN_COMPRESSION_PREFIX + NUM_COMPRESSION_PREFIXES;
+
+#ifdef _PARALLEL_SORT
+static constexpr bool USE_PARALLEL_SORT = true;
+#else
+static constexpr bool USE_PARALLEL_SORT = false;
+#endif
+static constexpr size_t NUM_SORT_THREADS = 4;