This is not yet finished at all, but I want to make it public for evaluation