@joka921 2019-03-26
@niklas88 @floriankramer I have fixed some bugs in the IdTable operators, please have a thorough look at this since you were the original creators and reviewers of this structure and such a basic data structure should be thoroughly tested and reviewed