@niklas88 2019-03-26
Ok good news, it turns out this does indeed give a measurable speedup. While it makes fast queries a little slower it also makes slow queries quite a bit faster (up to a factor of >2). Maybe we can still alleviate the former by tuning when we use which.

On Wikidata Full

With parallel sorting:

    Average completion moves: 4.52 (vs 8.51 manually)
    Executed 265 queries in 13608.614 seconds
    with the following timings

    Average: 51.35 ms/query
    Max: 729.06 ms/query
    90th percentile: 91.39 ms/query
    70th percentile: 72.32 ms/query
    50th percentile: 11.17 ms/query

Without parallel sorting:

    Average completion moves: 4.52 (vs 8.51 manually)
    Executed 272 queries in 15101.894 seconds
    with the following timings

    Average: 55.52 ms/query
    Max: 841.29 ms/query
    90th percentile: 100.72 ms/query
    70th percentile: 72.84 ms/query
    50th percentile: 8.65 ms/query

On Freebase

With parallel sorting:

    Average completion moves: 5.13 (vs 15.70 manually)
    Executed 18 queries in 21779.729 seconds
    with the following timings

    Average: 1209.98 ms/query
    Max: 16579.60 ms/query
    90th percentile: 2140.94 ms/query
    70th percentile: 225.67 ms/query
    50th percentile: 137.89 ms/query

Without parallel sorting:

    Average completion moves: 5.13 (vs 15.70 manually)
    Executed 18 queries in 45027.226 seconds
    with the following timings

    Average: 2501.51 ms/query
    Max: 38811.00 ms/query
    90th percentile: 2115.76 ms/query
    70th percentile: 397.80 ms/query
    50th percentile: 169.46 ms/query