@niklas88 2019-03-26
Hmm, something still seems buggy here. I'm seeing some weirdly aborted queries in the cache and server crashes @joka921 could you split this into 2 PRs, one with the `IdTable` fixes and one with the parallel sorting?