@joka921 2019-03-26
Ok,
1. I can split this up, but can I get some more information about the crashes (can you pin them down to "crashing during sort", "crashing somewhere else" and the static or dynamic idTable case?
Maybe we can debug this together tomorrow.