@niklas88 2019-03-26
@joka921 they are quite erratic and currently I can't even reproduce them. Tomorrow I'll only be at work till 15:30 at the latest but sure.