@niklas88 2019-03-29
Merged this by CLI so I could rebase away the clang format and debug commits.